% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcaliDocument.R
\name{ExcaliDocument}
\alias{ExcaliDocument}
\title{ExcaliDocument}
\description{
R6 Methods for ExcaliDocument object. Object structure is based on the
Excalidraw JSON Schema:
\url{https://docs.excalidraw.com/docs/codebase/json-schema}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{(\code{character(1)})\cr
The type of the Excalidraw schema}

\item{\code{version}}{(\code{integer(1)})\cr
The version of the Excalidraw schema}

\item{\code{source}}{(\code{character(1)})\cr
The source URL of the Excalidraw application}

\item{\code{elements}}{(\code{list()})\cr
A list of objects representing excalidraw elements on canvas}

\item{\code{appState}}{(\code{list()})\cr
Additional application state/configuration}

\item{\code{files}}{(\code{list()})\cr
List of data for excalidraw image elements}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExcaliDocument-new}{\code{ExcaliDocument$new()}}
\item \href{#method-ExcaliDocument-add}{\code{ExcaliDocument$add()}}
\item \href{#method-ExcaliDocument-replace}{\code{ExcaliDocument$replace()}}
\item \href{#method-ExcaliDocument-export}{\code{ExcaliDocument$export()}}
\item \href{#method-ExcaliDocument-clone}{\code{ExcaliDocument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ExcalidrawDoc object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-add"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-add}{}}}
\subsection{Method \code{add()}}{
Add a glyph to the scene
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{glyph attributes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-replace"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-replace}{}}}
\subsection{Method \code{replace()}}{
Replace all diagram components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$replace(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{List of replacement components}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-export"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-export}{}}}
\subsection{Method \code{export()}}{
Save the document as json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$export(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Export file path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
