% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minixcali-package.R
\docType{package}
\name{minixcali-package}
\alias{minixcali}
\alias{minixcali-package}
\title{minixcali: Excalidraw Document Builder}
\description{
Build Excalidraw documents with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baptiste/minixcali}
  \item \url{https://baptiste.github.io/minixcali/}
  \item Report bugs at \url{https://github.com/baptiste/minixcali/issues}
}

}
\author{
\strong{Maintainer}: Baptiste Auguie \email{baptiste.auguie@gmail.com}

}
\keyword{internal}
