% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{xkd_draw}
\alias{xkd_draw}
\title{xkd_draw}
\usage{
xkd_draw(
  x = 0,
  y = 0,
  width = 100,
  height = 100,
  angle = 0L,
  strokeColor = "#000000",
  backgroundColor = "transparent",
  fillStyle = c("solid", "hachure", "cross-hatch"),
  strokeWidth = 1L,
  strokeStyle = c("solid", "dashed", "dotted"),
  roughness = 0L,
  opacity = 100L,
  groupIds = list(),
  strokeSharpness = "round",
  isDeleted = FALSE,
  boundElementIds = NULL,
  points = list(list(0L, 0L), list(100, 0L)),
  lastCommittedPoint = NULL,
  startBinding = NULL,
  endBinding = NULL,
  startArrowhead = NULL,
  endArrowhead = NULL,
  version = 1,
  versionNonce = 12345
)
}
\arguments{
\item{x}{x position of glyph in px}

\item{y}{y position of glyph in px}

\item{width}{glyph width in px}

\item{height}{glyph height in px}

\item{angle}{in radians}

\item{strokeColor}{HTML colour code}

\item{backgroundColor}{HTML colour code}

\item{fillStyle}{"solid" or "hachure" or "cross-hatch"}

\item{strokeWidth}{1 or 2 or 4}

\item{strokeStyle}{"solid" or "dashed" or "dotted"}

\item{roughness}{degree of wiggliness, 0 or 1 or 2}

\item{opacity}{0 to 100}

\item{groupIds}{list of strings indicating groupings}

\item{strokeSharpness}{"sharp" or "round"}

\item{isDeleted}{tag for collaboration}

\item{boundElementIds}{list of IDs used for connectors}

\item{points}{of glyph}

\item{lastCommittedPoint}{of glyph}

\item{startBinding}{of glyph}

\item{endBinding}{of glyph}

\item{startArrowhead}{of glyph}

\item{endArrowhead}{of glyph}

\item{version}{of glyph for collaboration}

\item{versionNonce}{of glyph for collaboration}
}
\description{
xkd_draw
}
