% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{xkd_math}
\alias{xkd_math}
\title{xkd_math}
\usage{
xkd_math(
  x = 0,
  y = 0,
  width = 100,
  height = 100,
  angle = 0,
  strokeColor = "#000000",
  backgroundColor = "#868e96",
  fillStyle = "solid",
  strokeWidth = 2L,
  strokeStyle = "solid",
  roughness = 0L,
  opacity = 100L,
  groupIds = list(),
  strokeSharpness = c("sharp", "round"),
  isDeleted = FALSE,
  boundElementIds = NA,
  text = "$$\\\\nabla\\\\cdot\\\\mathbf{E}$$",
  fontSize = 36L,
  fontFamily = 1L,
  textAlign = "left",
  verticalAlign = "top",
  baseline = 32L,
  version = 300,
  versionNonce = 12345,
  subtype = "math",
  originalText = "$$\\\\nabla\\\\cdot\\\\mathbf{E}$$",
  useTex = TRUE
)
}
\arguments{
\item{x}{x position of glyph in px}

\item{y}{y position of glyph in px}

\item{width}{glyph width in px}

\item{height}{glyph height in px}

\item{angle}{in radians}

\item{strokeColor}{HTML colour code}

\item{backgroundColor}{HTML colour code}

\item{fillStyle}{"solid" or "hachure" or "cross-hatch"}

\item{strokeWidth}{1 or 2 or 4}

\item{strokeStyle}{"solid" or "dashed" or "dotted"}

\item{roughness}{degree of wiggliness, 0 or 1 or 2}

\item{opacity}{0 to 100}

\item{groupIds}{list of strings indicating groupings}

\item{strokeSharpness}{"sharp" or "round"}

\item{isDeleted}{tag for collaboration}

\item{boundElementIds}{list of IDs used for connectors}

\item{text}{text string}

\item{fontSize}{in pts}

\item{fontFamily}{1, 2, 3 (FG_Virgil,Cascadia,?)}

\item{textAlign}{"left", "center", "right"}

\item{verticalAlign}{"bottom", "center", "top"}

\item{baseline}{offset in pts?}

\item{version}{of glyph for collaboration}

\item{versionNonce}{of glyph for collaboration}
}
\description{
xkd_math
}
