% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{xkd_read}
\alias{xkd_read}
\title{Read an Excalidraw file}
\usage{
xkd_read(file, excalidoc = FALSE, simplifyVector = FALSE)
}
\arguments{
\item{file}{A file path or URL for an Excalidraw file with a JSON or
excalidraw file extension.}

\item{excalidoc}{Default \code{FALSE}. If \code{TRUE}, return a R6 \code{ExcaliDoc} class
object.}

\item{simplifyVector}{simplifies nested lists into vectors and data frames. See \code{\link[jsonlite:fromJSON]{fromJSON()}}.}
}
\value{
A named list or ExcaliDoc object
}
\description{
Read an Excalidraw file from a file or URL. If \code{simplifyVector = TRUE}, the
the diagram \code{elements} are returned as a data frame instead of a list. This
option makes the returned object is incompatible with the R6 methods in this
package.
}
