% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcaliDocument.R
\name{ExcaliDocument}
\alias{ExcaliDocument}
\title{ExcaliDocument}
\description{
R6 Methods for ExcaliDocument object
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{(\code{character(1)})\cr
Object type}

\item{\code{version}}{(\code{integer(1)})\cr
ExcaliDraw version number}

\item{\code{source}}{(\code{integer(1)})\cr
ExcaliDraw source}

\item{\code{appState}}{(\code{list()})\cr
ExcaliDraw app state}

\item{\code{files}}{(\code{list()})\cr
ExcaliDraw files}

\item{\code{elements}}{(\code{list()})\cr
ExcaliDraw elements}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExcaliDocument-new}{\code{ExcaliDocument$new()}}
\item \href{#method-ExcaliDocument-add}{\code{ExcaliDocument$add()}}
\item \href{#method-ExcaliDocument-export}{\code{ExcaliDocument$export()}}
\item \href{#method-ExcaliDocument-clone}{\code{ExcaliDocument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ExcalidrawDoc object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-add"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-add}{}}}
\subsection{Method \code{add()}}{
Add a glyph to the scene
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{glyph attributes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-export"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-export}{}}}
\subsection{Method \code{export()}}{
Save the document as json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$export(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Export file path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExcaliDocument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExcaliDocument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExcaliDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
